refactor: 👷 tests /interfaces to /typeDefinitions
